backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
motor==3.3.1
pymongo==4.6.1
python-magic==0.4.27
google-cloud-vision==3.4.4
python-jose==3.3.0
pydantic==2.4.2
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pytest>=7.4.3
httpx>=0.25.1
Pillow>=10.0.0 